Create Test Artist Page HTML
The HTML and CSS are in the same file because I can't work out how to add to add a folder with everything in it.
The logo is the inverted one. The image is The Camerata Singers.

Everything here can be changed, and as I was working without the design we want, I don't expect to keep much of it. I just wanted a small backbone.